+++ crates/module-system/module-implementations/sov-attester-incentives/tests/integration/challenger.rs
@@ -2,20 +2,24 @@ use std::convert::Infallible;
 use std::sync::atomic::AtomicU64;
 use std::sync::Arc;
 
-use sov_attester_incentives::{CallMessage, Event, SlashingReason};
+use sov_attester_incentives::{AttesterIncentives, CallMessage, Event, SlashingReason};
 use sov_bank::Amount;
+use sov_mock_da::MockDaSpec;
 use sov_modules_api::prelude::UnwrapInfallible;
+use sov_state::jmt::RootHash;
+use sov_state::StorageRoot;
 use sov_test_utils::generators::attester_incentive::framework::TestChallengeGenerator;
-use sov_test_utils::generators::attester_incentive::{
-    TestAttestationMessageError, TestChallengeMessageError,
-};
+use sov_test_utils::generators::attester_incentive::TestChallengeMessageError;
 use sov_test_utils::runtime::TestRunner;
 use sov_test_utils::{
-    AsUser, BondedTestChallenger, TestAttester, TransactionTestCase, TEST_DEFAULT_USER_STAKE,
-    TEST_ROLLUP_FINALITY_PERIOD,
+    AsUser, BondedTestChallenger, ProofTestCase, ProofType, TestAttester, TransactionTestCase,
+    TEST_DEFAULT_USER_STAKE, TEST_ROLLUP_FINALITY_PERIOD,
 };
 
-use crate::helpers::{setup, TestAttesterIncentives, TestRuntimeEvent, RT, S};
+use crate::helpers::{
+    build_challenge, build_proof, make_attestation_blob, make_challenge_blob, setup,
+    TestAttesterIncentives, TestRuntimeEvent, RT, S,
+};
 
 /// Helper that sets up a configuration where:
 /// - the challenger is bonded and
@@ -26,7 +30,7 @@ fn setup_with_wrong_attestation() -> (
     BondedTestChallenger<S>,
     Amount,
 ) {
-    let (mut runner, mut genesis_attester, mut genesis_challenger, _) = setup();
+    let (mut runner, genesis_attester, mut genesis_challenger, _) = setup();
 
     let genesis_attester_address = genesis_attester.user_info.address();
     let genesis_attester_bond = genesis_attester.bond;
@@ -38,8 +42,6 @@ fn setup_with_wrong_attestation() -> (
         Arc::new(AtomicU64::new(genesis_challenger.user_info.balance()));
     let expected_challenger_balance_2 = expected_challenger_balance.clone();
     let expected_challenger_balance_3 = expected_challenger_balance.clone();
-
-    let expected_attester_balance = genesis_attester.user_info.balance();
 
     let bond_challenger = TransactionTestCase {
         input: genesis_challenger.create_plain_message::<TestAttesterIncentives>(
@@ -81,42 +83,51 @@ fn setup_with_wrong_attestation() -> (
     let bonded_challenger =
         BondedTestChallenger::from_challenger(genesis_challenger, genesis_challenger_bond);
 
-    runner.execute_transaction(TransactionTestCase {
-        input: genesis_attester
-            .test_process_attestation(Err(TestAttestationMessageError::InvalidPostStateRoot)),
-        assert: Box::new(move |result, state| {
-            // Check that the working set has emitted a slashed event
-            assert!(result.events.iter().any(|event| matches!(
-                event,
-                TestRuntimeEvent::attester_incentives(Event::UserSlashed { .. })
-            )));
-            // Check that the attester was slashed
-            assert_eq!(
-                TestAttesterIncentives::default()
-                    .get_attester_bond_amount(&genesis_attester_address, state)
-                    .unwrap_infallible()
-                    .value,
-                0,
-            );
-            // Check that the transition was added to the challengeable set
-            assert_eq!(
-                TestAttesterIncentives::default()
-                    .bad_transition_pool
-                    // The attestation is added to the challengeable set at the slot of the attestation.
-                    // Since the attestation is invalid, the slot to attest is the same as the slot of the faulty attestation.
-                    // (the slot to attest is not increaed when the attestation is faulty)
-                    .get(&(1), state)
-                    .unwrap_infallible(),
-                Some(genesis_attester_bond),
-                "The failed attestation should have been added to the challengeable set"
-            );
-            assert_eq!(
-                TestRunner::<RT, S>::bank_gas_balance(&genesis_attester_address, state),
-                Some(expected_attester_balance - result.gas_used),
-                "The attester should have the correct bond amount from genesis"
-            );
-        }),
-    });
+    {
+        let mut attestation_proof = runner
+            .query_state(|state| build_proof(state, 1, &genesis_attester_address))
+            .unwrap();
+
+        attestation_proof.post_state_root =
+            StorageRoot::new(RootHash([255; 32]), RootHash([255; 32]));
+
+        runner.execute_proof::<TestAttesterIncentives>(ProofTestCase {
+            input: ProofType::Inline(make_attestation_blob(attestation_proof)),
+            override_sequencer: None,
+            assert: Box::new(move |_result, state| {
+                // TODO #1292:
+                // assert_matches!(
+                //    result.outcome.unwrap().outcome,
+                //    ProofOutcome::Invalid(InvalidProofError::PreconditionNotMet(_))
+                // );
+
+                // Check that the attester was slashed
+                assert!(TestAttesterIncentives::default()
+                    .bonded_attesters
+                    .get(&genesis_attester_address, state)
+                    .unwrap()
+                    .is_none(),);
+
+                // Check that the transition was added to the challengeable set
+                // The attestation should be part of the challengeable set and its associated value should be the BOND_AMOUNT
+                assert_eq!(
+                    AttesterIncentives::<S, MockDaSpec>::default()
+                        .bad_transition_pool
+                        .get(&1, state)
+                        .unwrap_infallible(),
+                    Some(genesis_attester_bond),
+                    "The transition should exist in the pool"
+                );
+
+                // TODO #1292:
+                // assert_eq!(
+                //    TestRunner::<RT, S>::bank_gas_balance(&genesis_attester_address, state),
+                //    Some(expected_attester_balance - result.gas_used),
+                //    "The attester should have the correct bond amount from genesis"
+                // );
+            }),
+        });
+    }
 
     (
         runner,
@@ -128,18 +139,21 @@ fn setup_with_wrong_attestation() -> (
 
 /// Test that given an invalid transition, a challenger can successfully challenge it and get rewarded
 /// This tests the happy path of challenge processing.
-// TODO: #1262
-#[ignore]
+
 #[test]
 fn test_valid_challenge() -> Result<(), Infallible> {
-    let (mut runner, _, bonded_challenger, expected_reward) = setup_with_wrong_attestation();
+    let (mut runner, _, bonded_challenger, _expected_reward) = setup_with_wrong_attestation();
     let bonded_challenger_address = bonded_challenger.user_info.address();
-    let mut bonded_challenger_balance = bonded_challenger.user_info.balance();
+    let _bonded_challenger_balance = bonded_challenger.user_info.balance();
 
-    // Then challenge the wrongly attested slot.
-    runner.execute_transaction(TransactionTestCase {
-        input: bonded_challenger.test_process_challenge_at_slot(Ok(()), 1),
-        assert: Box::new(move |result, state| {
+    let challenge_proof = runner
+        .query_state(|state| build_challenge(state, 1, bonded_challenger_address))
+        .unwrap();
+
+    runner.execute_proof::<TestAttesterIncentives>(ProofTestCase {
+        input: ProofType::Inline(make_challenge_blob(challenge_proof, true, 1)),
+        override_sequencer: None,
+        assert: Box::new(move |_result, state| {
             assert_eq!(
                 TestAttesterIncentives::default()
                     .bad_transition_pool
@@ -149,24 +163,18 @@ fn test_valid_challenge() -> Result<(), Infallible> {
                 "The transition should have disappeared from the pool"
             );
 
-            // Check that a reward event has been emitted
-            assert!(result.events.iter().any(|event| matches!(
-                event,
-                TestRuntimeEvent::attester_incentives(
-                    Event::ProcessedValidProof { challenger }) if *challenger == bonded_challenger_address
-                )
-            ));
+            // TODO #1292: check reward
 
-            bonded_challenger_balance -= result.gas_used;
-            let reward = TestAttesterIncentives::default()
-                .burn_rate()
-                .apply(expected_reward);
-
-            assert_eq!(
-                TestRunner::<RT, S>::bank_gas_balance(&bonded_challenger_address, state),
-                Some(bonded_challenger_balance + reward),
-                "The challenger has not been rewarded the correct amount"
-            );
+            // TODO #1292:
+            // bonded_challenger_balance -= result.gas_used;
+            // let reward = TestAttesterIncentives::default()
+            //    .burn_rate()
+            //    .apply(expected_reward);
+            //assert_eq!(
+            //    TestRunner::<RT, S>::bank_gas_balance(&bonded_challenger_address, state),
+            //    Some(bonded_challenger_balance + reward),
+            //    "The challenger has not been rewarded the correct amount"
+            // );
         }),
     });
 
@@ -180,7 +188,7 @@ fn test_invalid_challenge_helper(
     let (mut runner, _, bonded_challenger, expected_reward) = setup_with_wrong_attestation();
 
     let bonded_challenger_address = bonded_challenger.user_info.address();
-    let bonded_challenger_balance = bonded_challenger.user_info.balance();
+    let _bonded_challenger_balance = bonded_challenger.user_info.balance();
 
     // Then challenge the wrongly attested slot.
     runner.execute_transaction(TransactionTestCase {
@@ -216,12 +224,14 @@ fn test_invalid_challenge_helper(
                 Some(expected_reward),
                 "The transition should *not* have disappeared from the pool"
             );
+
             // Check that the challenger was not rewarded
-            assert_eq!(
-                TestRunner::<RT, S>::bank_gas_balance(&bonded_challenger_address, state),
-                Some(bonded_challenger_balance - result.gas_used),
-                "The challenger balance is not correct"
-            );
+            // TODO: #1262
+            // assert_eq!(
+            //    TestRunner::<RT, S>::bank_gas_balance(&bonded_challenger_address, state),
+            //    Some(bonded_challenger_balance - result.gas_used),
+            //    "The challenger balance is not correct"
+            //);
         }),
     });
 }