+++ crates/module-system/sov-modules-api/src/proof_metadata.rs
@@ -64,20 +64,14 @@ impl<S: Spec> ProofSerializer for SovApiProofSerializer<S> {
         &self,
         serialized_proof: SerializedAggregatedProof,
     ) -> anyhow::Result<Vec<u8>> {
-        let details = TxDetails::<S> {
-            max_priority_fee_bips: PriorityFeeBips::ZERO,
-            max_fee: 10_000_000,
-            gas_limit: None,
-            chain_id: config_value!("CHAIN_ID"),
-        };
+        let details: TxDetails<S> = make_details(10_000_000);
 
         let proof_with_details = SerializeProofWithDetails {
             proof: ProofType::ZkAggregatedProof(serialized_proof),
             details,
         };
 
-        // TODO: Put SerializedAggregatedProof directly on chain without wrapping in a vec <https://github.com/Sovereign-Labs/sovereign-sdk-wip/issues/1065>
-        let serialized_proof_with_details = borsh::to_vec(&borsh::to_vec(&proof_with_details)?)?;
+        let serialized_proof_with_details = serialize_proof_with_details(&proof_with_details)?;
 
         Ok(serialized_proof_with_details)
     }
@@ -86,21 +80,48 @@ impl<S: Spec> ProofSerializer for SovApiProofSerializer<S> {
         &self,
         serialized_attestation: SerializedAttestation,
     ) -> anyhow::Result<Vec<u8>> {
-        let details = TxDetails::<S> {
-            max_priority_fee_bips: PriorityFeeBips::ZERO,
-            max_fee: 10_000_000,
-            gas_limit: None,
-            chain_id: config_value!("CHAIN_ID"),
-        };
+        let details: TxDetails<S> = make_details(10_000_000);
 
         let proof_with_details = SerializeProofWithDetails {
             proof: ProofType::OptimisticProofAttestation(serialized_attestation),
             details,
         };
 
-        // TODO: Put SerializedAggregatedProof directly on chain without wrapping in a vec <https://github.com/Sovereign-Labs/sovereign-sdk-wip/issues/1065>
-        let serialized_proof_with_details = borsh::to_vec(&borsh::to_vec(&proof_with_details)?)?;
+        let serialized_proof_with_details = serialize_proof_with_details(&proof_with_details)?;
 
         Ok(serialized_proof_with_details)
     }
+
+    fn serialize_challenge_blob_with_metadata(
+        &self,
+        serialized_challenge: SerializedChallenge,
+        slot_height: u64,
+    ) -> anyhow::Result<Vec<u8>> {
+        let details: TxDetails<S> = make_details(10_000_000);
+
+        let proof_with_details = SerializeProofWithDetails {
+            proof: ProofType::OptimisticProofChallenge(serialized_challenge, slot_height),
+            details,
+        };
+
+        let serialized_proof_with_details = serialize_proof_with_details(&proof_with_details)?;
+
+        Ok(serialized_proof_with_details)
+    }
+}
+
+fn make_details<S: Spec>(max_fee: u64) -> TxDetails<S> {
+    TxDetails {
+        max_priority_fee_bips: PriorityFeeBips::ZERO,
+        max_fee,
+        gas_limit: None,
+        chain_id: config_value!("CHAIN_ID"),
+    }
+}
+
+fn serialize_proof_with_details<S: Spec>(
+    proof_with_details: &SerializeProofWithDetails<S>,
+) -> anyhow::Result<Vec<u8>> {
+    // TODO: Put SerializedAggregatedProof directly on chain without wrapping in a vec <https://github.com/Sovereign-Labs/sovereign-sdk-wip/issues/1065>
+    Ok(borsh::to_vec(&borsh::to_vec(&proof_with_details)?)?)
 }